requirements.txt
torch>=2.1.0
torchvision>=0.10.0
torchaudio>=0.10.0
numpy>=1.24.0
pandas>=1.3.3
matplotlib>=3.4.3
Pillow>=8.3.2
opencv-python>=4.5.3.56
transformers>=4.11.3
ultralytics>=0.18.2
albumentations>=1.1.0